thrift: create git submodule for upstream thrift
Add a git submoudle to point to upstream thrift at
https://github.com/cfriedt/thrift.git

We will be making project-specific changes in the `zephyr` branch
of that project.

The checkout corresponds with `v0.16.0`, initially.

Placed in `.upstream` for lack of a better place.

IIRC, `west` is now configured to also clone submodules by default
so the README likely does not need to be updated.

Fixes #2

Signed-off-by: Chris Friedt <chrisfriedt@gmail.com>