test: thrift: hello: enable C++ exceptions
Most of the Thrift C++ runtime depends on having exceptions
enabled. Normally, for deeply embedded runtimes like Zephyr
this is something we try to avoid. However, Zephyr does
support exceptions.

Somewhat tangential, but exceptions are baked into a large
portion of the C++ standard APIs, so it would be difficult
to pry that apart within Thrift itself.

In the long run, likely a newer standard C generator for
Thrift would be preferable for resource-constrained cases.

Fixes #21

Signed-off-by: Chris Friedt <chrisfriedt@gmail.com>