submanifest: use built-in support for git submodules
Previously, when the project was structured as a downstream
application, we needed to do some extra legwork to get the
top-level git submodules to be pulled in, because it is
not supported in west. Since then, we've switched to using
submanifests, and git submodules are directly supported.

No need to manually specify submodules as west repos.

Signed-off-by: Chris Friedt <cfriedt@meta.com>