cmake: use ZEPHYR_THRIFT_MODULE_DIR
It's useful to be able to reference files relative to the root of
the module. Previously, this was not working in CI because of some
weird git submodule / west repo hackery.

However, since we're not using a proper submanifest and git
submodule, this should work fine.

Noted, `ZEPHYR_CURRENT_MODULE_DIR` was the first choice but does
not work reliably (see chre and sof workarounds). The build system
automatically generates `ZEPHYR_THRIFT_MODULE_DIR` as well. If the
path ever changes it does need to be modified.

Signed-off-by: Chris Friedt <cfriedt@meta.com>