upstream: add .apache-thrift as new upstream git submodule
Remove the `.upstream` git submodule, which pointed to
`cfriedt/thrift` and add a new git submodule, `.apache-thrift`,
which points to `apache/thrift`.

This is significant, because it marks that milestone that all
required changes have been made to upstream thrift. The only
remaining changes left in this module are specifically to
address issues with Zephyr's C++ implementation. Specifically,
`std::thread`, `std::mutex`, `std::condition_variable` support,
and so on.

Signed-off-by: Chris Friedt <cfriedt@meta.com>