thrift: TConnectedClient: workaround for non-virtual dtor
Upstream has an issue in that `TConnectedClient` has a non-
virtual destructor. This results in a warning from GCC
when compiling `TServerFramework.cpp`. Of course, in CI,
all of Zephyr's warnings are promoted to errors, which
will break the build.

Tracking the issue upstream at the link below.

https://issues.apache.org/jira/browse/THRIFT-5678

We'll keep this workaround in tree until the fix is
merged upstream.

Fixes #160

Signed-off-by: Christopher Friedt <cfriedt@meta.com>