lib: thrift: TSSLServerSocket: close pointer to interrupt socket
It would seem that Zephyr's C++ runtime was not properly
executing custom deleters for `std::shared_ptr`. Originally,
it seemed as though `TServerSocket.cpp` was leaking file
descriptors, but there is a custom deleter used that is
never called. Details at the URL below.

https://issues.apache.org/jira/browse/THRIFT-5677

Once this is fixed in Zephyr's C++ runtime, this workaround
should be reverted.

Fixes #161

Signed-off-by: Christopher Friedt <cfriedt@meta.com>